Mms : Don't close single helper db instances
Starting 4.1, android uses a SQLiteConnectionPool. If you
obtain a re-used SQLiteDatabase instance and query from it,
if by the time you perform an action on the cursor someone closes
that SQLiteDatabase instance, you will get the
"Cannot perform this operation because the connection pool has been closed."
exception.

Prevent that by only closing the db if necessary.

Change-Id: I2f4adb38c6a12d73beaee9df54c7ad39d4fe8cd0